Added html to database and rendered the html
