--- database-mongo/included.js
@@ -1,164 +1,0 @@
-  module.exports.file = [{ included: [
-      "PCB",
-      "Steel plate",
-      "Anodized CNC aluminum case",
-      "USB cable",
-      "Choice of keycaps"
-    ],
-    prod_name: "planck-mechanical-keyboard",
-    boxContent: [`
-      YOU MADE IT HAPPEN!
-      This is one of the most requested options from the Most Wanted ~40% Keyboards poll created by tobsn. Thanks to everyone who voted for making this drop possible.
-      1 of 34 Options    3494 Votes`, 
-      `Note: At checkout, you’ll have your choice of the Mid-Pro or Hi-Pro case in black, green, purple, or space gray. You’ll also have your choice of a Matias-compatible plate or an MX-compatible plate (both are compatible with the case) as well as the option to add or remove keycaps, switches, and a protective sleeve. See checkout options for more details.
-
-      To allow us to start production on the order, and to ensure smooth delivery, we will be collecting payment on October 31. If this is an issue for you, please join the drop after this date.`],
-    descriptionHeader: [
-      "The New, More Powerful Planck", "Sturdy, Compatible Plate Design", "Strong Anodized Aluminum Case", "Fully Programmable PCB", "Beeps & Boops"
-    ],
-    description: [
-        "The ortholinear 40% Planck mechanical keyboard is hailed for its unique layout—and the new version is even better. Laid out so that every key is within two units of your fingers on the home row, this compact keyboard minimizes finger travel. The “raise” and “lower” space bar keys make it easy to access additional layers for extra functionality. For this round, the CNC aluminum case comes in four different colors, including the striking purple colorway offered on the Preonic. This version also has Kaihua hot-swappable switch sockets, so you can swap out your switches without ever picking up a soldering iron. What’s more, the new Planck has an upgraded ARM STM32 processor that provides more power to the PCB and USB-C connector—and it comes in two case styles and four colors.",
-      "Like the previous version, the new Planck comes with your choice of an MX plate (for MX switches and clones) and a Matias plate (for Matias and ALPS switches), both of which support 1 x 2u and Grid layout. However, this version features Kaihua hotswap sockets, which support grid (2x1u), 1x2u, and 2x2u bottom row layouts. Made from 304 stainless steel, the plate is heavy and stable, with five holes for mounting the included M2 screws.", 
-      "On the old Planck, you could see the plate atop the milled aluminum body. For this version, we’ve tucked the plate into the case for a more streamlined look—one you can get two different ways, depending on your preference. The milled bottom is made from a single piece of aluminum with thick 3.1-millimeter walls. It’s anodized with a smooth finish in your choice of four colors. A hole near the USB port allows easy access to the reset button on the PCB. The kit has all the hardware you need to assemble the keyboard, plus four 3M rubber feet so it won’t slide around on your desktop.", 
-      "The matte black PCB features an AVR chip and can be programmed however you like. It comes pre-programmed with the standard layout, but it also has a reset button to reprogram the keyboard. It’s compatible with Cherry MX and Matias/ALPS switch footprints, and all components of the board have been pre-soldered so the PCB is ready for key switches and LEDs. Plus, the hotswap sockets on this version allow for easy kit building when you assemble your board. (Please note that the Matias switches will need to be soldered onto the PCB. Only MX-style switches—Gateron, Kaihua, Cherry MX, etc.—will be able to use the Kaihua hot swap sockets.)",
-      "Also notable, the PCB is outfitted with a small speaker, which is driven by the MCU. In the default firmware this will make a few different noises—like a short series of beeps on startup, notifications when you put the board into Device Firmware Upgrade (DFU) mode, plus noises when you change the default layout to prevent mishaps when typing quickly. The speaker can be disabled by turning off the audio via a keycode or by flashing new firmware."
-    ],
-    imageHeader: ["Case Options","Keycap Options","Switch options"],
-    images: ["https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/IhT3NhxzSAONgRhKbY7a_AI7B5534-copy.jpg?auto=format&fm=jpg&fit=crop&w=924&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/cxWI1GyoSGGch6osWQrP_AI7B6777-copy.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/ZOr1IE4RxuDmTVjv0gsJ_AI7B6804-copy.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/XbWiNmBkQPqwVhk8EEMa_AI7B5416-copy-checkout1.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/Ew9RSUodQaCBRJ3KL1xM_AI7B5416-copy-checkout2.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/F9Q4Ma2YR0WVVKlLbu3C_AI7B5192-copy-pc-dark.jpg?auto=format&fm=jpg&fit=crop&w=924&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/a7tb3p9fQHiqFCtDzL3z_AI7B5383-copy-dark.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/Vu4OfnOCTa46XPDHUfvw_AI7B5392-copy.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/5o6hhUzkQrCN0g4AC6la_AI7B5218-copy-dark.jpg?auto=format&fm=jpg&fit=crop&w=924&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/OIH3lVQ5SNqStf2RXnSr_planck-rev6-pcb-top-align.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/4EWffL7GSfiXTtlf7u5Z_planck-rev6-hot-swap-top-trans.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/T2HlEoTFSh2kbj4ObQqH_AI7B5216-copy-dark.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/QAdQbuLIQIKWEZDK4Ab4_AI7B5230-copy-dark.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/YeMbU6nQt6XfDAzDBgBj_AI7B6031-copy.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/ms5pJ4yR2558gkDFetCA_AI7B5498-copy.jpg?auto=format&fm=jpg&fit=crop&w=458&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/gZBQEopFTvSpwd80t9sS_highprofile-purple.jpg?auto=format&fm=jpg&fit=crop&w=490&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/FP/2VFMahX0QAa18SD4Wpou_acute_keycap_set_planck_rev6.jpg?auto=format&fm=jpg&fit=crop&w=490&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/planck-mechanical-keyboard/MD-28284_20170309113239_dfbc5f3a067f58d3.jpg?auto=format&fm=jpg&fit=crop&w=490&bg=f0f0f0&dpr=1"],
-    video: "https://youtu.be/bEPg8kk84gw", 
-    shippingDate: "Apr 30, 2019 PT"
-  },
-  {
-    included:[
-      "Plastic case",
-      "Spare O-rings",
-      "Key ring for attachment",
-    ],
-    prod_name: "massdrop-copper-aaa-pocket-flashlight",
-    descriptionHeader:["An EDC Favorite With a Brand-New Copper Body", "Additional Features",],
-    description: ["Our third version of the Massdrop AAA flashlight, this model features a hefty copper body that will develop a natural patina over time. About the size of a tube of chapstick, it pumps out up to 110 lumens and casts a beam as far as 154 feet. The titanium button allows you to cycle from moonlight mode (great for not disrupting others in tight spaces) to medium mode to high mode. Perhaps best of all, it runs on a single AAA battery for a whopping 36 hours on the lowest setting. At checkout, you’ll have your choice between the CREE XP-G2 R5 LED, which casts a brighter white beam, and the brand-new Nichia 219C LED. Based on community preferences from the last run, the Nichia 219C LED is more efficient and approximately 6 percent brighter than the Nichia 219B, with only a modest amount of warmth lost.", 
-    "The flashlight’s copper body features knurling for grip in slippery conditions. And if you do happen to drop it in the water, not to worry: It’s IPX-8 waterproof and 1.5-meter impact resistant. Attach it to your belt or pocket, reverse the clip and put it on the bill of a cap, or remove the clip altogether and slide the light into a pocket. The stainless steel clip is a nice color complement to the glistening orange body. It also stands on its tail to shed low light in a whole room."
-  ],
-    boxContent: "Note: This item is in stock and will ship within 1 business day with standard shipping. Your payment method will be charged 2 hours after joining. Cancellations are accepted until that time. The drop will launch and end each day to collect purchases. At checkout, choose between the XP-G2 R5 LED and the Nichia 219C LED. ",
-    images: ["https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/AI7B7826_copy_final_silver_20180118172316.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/AI7B7719_copy_silver_20180118172935.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/slack-imgs_20180124161025.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/lights_3_20180124162153.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/copper_20180124154141.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/AI7B7904_copy_silver_20180118172313.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/AI7B7733_copy_silver_20180118172313.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/AI7B9552_copy_page_20180123163423.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-copper-aaa-pocket-flashlight/Untitled2_copy_dontresize_new_page_silver_20180118172314.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1"],
-    video: "https://vimeo.com/252412492",
-    shippingDate: "Oct 26, 2018 PT",
-    specs:[
-      "Material: Copper",
-      "Titanium button",
-      "Mode sequence: L-M-H",
-      "Battery: 1 x AAA",
-      "Removable, reversible stainless steel pocket clip",
-      "IPX-8 waterproof",
-      "1.5 m impact resistant",
-      "Working voltage: 0.9 – 1.5V",
-      "Dimensions: 2.9 x 0.6 in (7.4 x 1.5 cm)",
-      "Weight (with battery): 1.5 oz (42.5 g)",
-      "CREE XP-G2 R5 LED",
-      
-      "Max beam intensity: 553 cd",
-      "Max beam distance: 154 ft (47 m)",
-      "Low mode: 5 lumens, 36 hrs",
-      "Medium mode: 32 lumens, 4 hrs",
-      "High mode: 110 lumens, 30 min",
-      "Nichia 219C LED",
-      
-      "Max beam intensity: 306 cd",
-      "Max beam distance: 114.8 ft (35 m)",
-      "Low mode: 3 lumens",
-      "Medium mode: 20 lumens",
-      "High mode: 85 lumens"
-    ]
-    // <iframe src="//player.vimeo.com/video/252412492?title=0&amp;byline=0&amp;portrait=0" allowfullscreen="" frameborder="0"></iframe>
-  },
-  {
-    included: [
-      "Carrying pouch",
-      "2-year warranty"
-    ],
-    prod_name: "massdrop-fostex-tr-x00-purpleheart-headphones",
-    descriptionHeader:[
-      "A Community Favorite, Now With Removable Cables",
-      "A Side-by-Side Look",
-      "Intimate Sound",
-      "Comfort-First Design",
-      "Community Reviews for the TH-X00 Purpleheart"
-    ],
-    description: [
-      "When the Massdrop community asked for a closed-back headphone that redefined price-to-performance expectations, the TH-X00 was our answer. Made in collaboration with Fostex, a brand that’s produced innovative audio gear since 1949, these headphones feature purpleheart ear cups, a magnesium alloy frame, and Fostex’s signature biodynamic drivers. Purpleheart is among the densest woods in the world, and along with adding a darker, more sophisticated aesthetic, it delivers a slightly different sound signature. Now, in another response to the community’s requests, we’ve updated the original with a removable cable. Introducing the TR-X00 (the “R” stands for “removable”): everything you loved about the TH-X00, now with a detachable cable for greater flexibility. ",
-      "The density of purpleheart wood makes it excellent for reducing acoustic resonance. As a result, the sub-bass through bass is increased ~2 dB, while following the same curve as the original, which peaks at 50 Hz. The lower midrange is decreased by ~5 dB, and while the mahogany version had a slight peak at ~600 Hz, the purpleheart has a smoother climb from 500 to 1000 Hz. The midrange proper is essentially identical to that of the mahogany, the upper midrange is ~2 dB less throughout, and the highs are slightly more emphasized. Peaks and valleys are positioned differently than on the mahogany, but their magnitude is similar. As for the super highs, while the mahogany falls off at ~14 kHz, the purpleheart maintains presence through that range. Finally, CSD graphs of the mahogany showed some resonance at 3,000 and 6,000 Hz; these are all but eliminated in the purpleheart.",
-      "Equipped with the biodynamic cone Fostex first developed in the late ‘80s, which delivers a sound unique to any other driver structure, the TR-X00 benefits from years of iteration on the initial design. The higher magnetic flux density in these headphones (more than 1 tesla) creates instant attack and decay for a crisper, more detailed sound signature. With the emphasis on the low end toned down slightly from the TH-600, the rich bass meets the Fostex standard without becoming bloated. The result is an incredibly intimate ",
-      "The TR-X00’s wood ear cups are a nod to the Denon D series and complemented by Fostex’s proprietary magnesium alloy body, built to be lightweight, strong, and durable. Made of eggshell protein, the ear pads’ soft leatherette is designed to form a super tight seal, while still being breathable for hours of comfortable listening. And because the curve of the headband has been optimized to fit the human head, minimal padding is needed. The thick, braided cable has a conveniently long distance between the headphones and Y split. The Massdrop logo is silk-screened on the inside of the headband, just above the left cup, with a serial number located just above the right cup. ",
-      "“If it sounds better than the R-10, it has to be good. It blew my mind, got all three of them. Different flavors, different look, and sound like no other.” – Massdrop member albertofgomez"
-    ],
-      boxContent: `Note: This item is in stock and will ship within 1 business day. Your payment method will be charged 2 hours after joining. Cancellations are accepted until that time. The drop will launch and end each day to collect purchases.
-
-
-
-      Each pair of headphones is one of a kind, and due to the nature of wood, slight variations in wood grain may occur. If you’re an international customer looking to partner with a shipping forwarder, please read this FAQ.`,
-    shippingDate: "Oct 26, 2018 PT",
-    specs: ["Fostex x Massdrop",
-      "Closed-back design",
-      "50mm dynamic transducer",
-      "Magnesium alloy construction",
-      "Purpleheart wood ear cups, brilliant gloss finish",
-      "Leatherette ear pads, matte black",
-      "Magnetic flux density: >1 tesla",
-      "Impedance: 25 ohms",
-      "Sensitivity: 94 dB/mW",
-      "Maximum input: 1,800 mW",
-      "Frequency response: 5–45,000 Hz",
-      "Removable 10 ft (3 m) thick, braided Y cable",
-      "1/4 in (6.3 mm) gold-coated stereo phone plug",
-      "Weight without cable: 13.6 oz (385 g)",
-      "Weight with cable: 18.2 oz (517 g)",
-      "CREE XP-G2 R5 LED",
-      "",
-      "Max beam intensity: 553 cd",
-      "Max beam distance: 154 ft (47 m)",
-      "Low mode: 5 lumens, 36 hrs",
-      "Medium mode: 32 lumens, 4 hrs",
-      "High mode: 110 lumens, 30 min",
-      "Nichia 219C LED",
-      "",
-      "Max beam intensity: 306 cd",
-      "Max beam distance: 114.8 ft (35 m)",
-      "Low mode: 3 lumens",
-      "Medium mode: 20 lumens",
-      "High mode: 85 lumens"
-    ],
-      images:["https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/ulK15ut2SIq0O3ZRxj7r_AI7B5458%20copy.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/KZ7VCJ1wQFOguQebb0P4_AI7B3129%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/0Td5OouT9ieIZazb4RT1_AI7B1795%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/WdPvHjF4Qju02mPdH42t_AI7B3155%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/dOoIVblIR2ac6P66xjPv_AI7B3107%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/YzTkZiEhSXmGOSQ95X88_AI7B3189%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/p3ic4JTQRdqSJfC9wDTP_AI7B3163%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/anRIV9TSqdKOzK771Tzg_AI7B3104%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/zJRY6YhRVu2xalIEPHdw_AI7B1825%20copy.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/eWCgZGRvTuis7MVBhUZY_AI7B3222%20copy.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/massdrop-fostex-tr-x00-purpleheart-headphones/FP/aR2IQIYORLWD06KFTJqV_AI7B9664%20copy%20tint.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1"]
-  },
-  {
-    included: [
-      "Fingernail clippers",
-      "Toenail clippers",
-      "Nail file",
-      "Nail nipper",
-      "Nostril scissors",
-      "Extra-grip slant tweezers",
-      "Zippered leather case",
-    ],
-    details:[
-      "Seki Edge",
-      "Model number: SS-3103",
-      "6-piece set",
-      "Stainless steel tools",
-      "Leather case with elastic straps"
-    ],
-    shippingDate: "Nov 23, 2018 PT",
-    prod_name: "seki-edge-grooming-kit",
-    decriptionHeader: ["New, Improved Quality","Contents of the 6-Piece Set","About Seki Edge"],
-    description: [
-      "Now sourced directly from Seki Edge, this six-piece grooming kit features the same great essentials previously offered on Massdrop, but with the improved quality that comes from working directly with Seki. From the stainless steel and shaping of the blades to the finishing on the tools, the craftsmanship is second to none. And each piece is branded with the Seki name to prove it.", 
-      "For nail care, the kit comes with curved satin clippers—one pair for fingernails and another for toenails. It also comes with a nail file and, to help trim dead skin, a stainless steel nail nipper. For nose maintenance, the kit includes a pair of nostril scissors. And to finish your grooming regimen, this six-piece set features extra-grip slant tweezers for plucking errant hairs",
-      "Rich with history, the Japanese town of Seki City was, as far back as the 13th century, known for its samurai-swordmaking prowess. And while swords are no longer manufactured here, some of the country’s finest cutlery is. It’s this centuries-old craftsmanship that inspires Seki Edge to produce men’s grooming instruments that are as beautiful as they are durable and functional."
-    ],
-    boxContent: `
-      YOU MADE IT HAPPEN!
-      This is one of the most requested options from the Grooming Set poll created by a community member. Thanks to everyone who voted for making this drop possible.
-      1 of 3 Options    62 Votes`,
-    images: ["https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140122_55bc1e6f8abdab8e.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140121_6822ac4fc567f904.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140121_cf0b2f9b40987d9c.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1","https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140121_a7a719bbdc5820cf.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140119_6ae07590600c9eb1.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140116_ae1e49ea4ed70295.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140120_9530bb0e7d7210ed.jpg?auto=format&fm=jpg&fit=crop&w=473&bg=f0f0f0&dpr=1", "https://massdrop-s3.imgix.net/product-images/seki-edge-grooming-kit/MD-22295_20160630140124_2b713cd4c2a512bf.jpg?auto=format&fm=jpg&fit=crop&w=955&bg=f0f0f0&dpr=1"]
-  }
-
-]
