+++ react-client/src/components/Gallery.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+const Gallery = (props) => {
+    return (
+        <img alt="Massdrop x Fostex TR-X00 Purpleheart Headphones" src="https://helpdesk.massdrop.com/hc/en-us/articles/216723438"/>
+    )
+  }
+  
+  export default Gallery;
